requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
astunparse==1.6.3
async-timeout==4.0.2
asyncpg==0.25.0
attrs==21.4.0
beautifulsoup4==4.10.0
braceexpand==0.1.7
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
frozenlist==1.3.0
idna==3.3
import-expression==1.1.4
importlib-metadata==4.11.3
numpy==1.22.3
psutil==5.9.0
python-dateutil==2.8.2
python-dotenv==0.19.2
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
urllib3==1.26.9
waifuim.py==3.2.1
yarl==1.7.2
zipp==3.7.0